Refactor EventBus gRPC transport to use `v1alpha` schema and improve server streaming call initialization.
Motivation:

- Transition to `v1alpha` schema for EventBus transport messages for better compatibility and clarity.
- Simplify and improve the initialization of server-side streaming calls with a more modular design.

Changes:

- Updated protobuf definitions and transport schema from `eventbus.transport` to `eventbus.transport.v1alpha`.
- Refactored `EventBusGrpcServerStreamingCall` to use a dedicated `init` method for inbound message consumer setup.
- Simplified server-side streaming call logic by removing `AtomicReference` and directly managing completion callbacks.
- Updated documentation to specify the new schema path and updated associated examples.
