+++ vertx-grpc-eventbus/src/main/java/io/vertx/grpc/eventbus/impl/EventBusGrpcCodec.java
@@ -10,7 +10,7 @@ import io.vertx.grpc.common.GrpcMessageDecoder;
 import io.vertx.grpc.common.GrpcMessageEncoder;
 import io.vertx.grpc.common.GrpcStatus;
 import io.vertx.grpc.common.WireFormat;
-import io.vertx.grpc.eventbus.transport.TransportFrame;
+import io.vertx.grpc.eventbus.transport.v1alpha.TransportFrame;
 
 final class EventBusGrpcCodec {
