--- vertx-grpc-eventbus/src/main/proto/eventbus_transport.proto
@@ -1,79 +1,0 @@
-syntax = "proto3";
-
-package io.vertx.grpc.eventbus;
-
-option java_package = "io.vertx.grpc.eventbus.transport";
-option java_multiple_files = true;
-
-// Wire protocol for streaming gRPC over the Vert.x event bus.
-//
-// Unary stays on the existing request/reply mapping (action header = method,
-// body = payload, single reply) and never uses these frames. It is wire
-// compatible with a Vert.x service proxy.
-//
-// Both sides know the call type from the ServiceMethod. A streaming call opens
-// with an event bus request() carrying the first message as its body, and the
-// server replies with an Ack giving its inbound address. From there both sides
-// exchange frames with fire-and-forget send() on a pair of addresses dedicated to
-// the call.
-//
-// The frame is serialized as protobuf binary (the event bus body is a Buffer).
-// The Message payload carries the bytes the gRPC encoder already produced for the
-// call's wire format, so it is not re-encoded here. Every variant is its own
-// message type so the oneof reads uniformly and each can grow fields later.
-message TransportFrame {
-  uint64 seq = 1; // per-stream, monotonic, advances on message frames only
-
-  oneof frame {
-    Ack ack = 2; // server -> client, the streaming open reply
-    Message message = 3; // a payload (encoder output), either direction
-    WindowUpdate window_update = 4; // flow-control credit, either direction
-    HalfClose half_close = 5; // client -> server, end of the request stream
-    Trailers trailers = 6; // server -> client, terminates the call
-    Cancel cancel = 7; // either direction, abnormal end
-    Headers headers = 8; // server -> client, response initial metadata, sent before the first message
-  }
-}
-
-// Server -> client, the reply to a streaming call's opening request(), giving the
-// server inbound address. Response initial metadata does not ride here: it is sent
-// later as a Headers frame, once the application handler has had a chance to set it.
-message Ack {
-  string server_address = 1; // server inbound stream address (client -> server frames)
-  uint32 initial_window = 2; // messages the server grants the client to send
-}
-
-// Server -> client, response initial metadata. The metadata itself rides as
-// __header__. prefixed delivery headers, as in unary. This frame is the in-order
-// marker that carries it ahead of the first message.
-message Headers {
-}
-
-// A payload, either direction. The bytes are the gRPC encoder's output for the
-// call's wire format, carried verbatim and handed straight to the decoder.
-message Message {
-  bytes payload = 1;
-}
-
-// Flow-control credit, either direction. The message-counted analog of an HTTP/2
-// WINDOW_UPDATE: grants the peer delta more messages to send.
-message WindowUpdate {
-  uint32 delta = 1;
-}
-
-// Client -> server, end of the request stream (half close).
-message HalfClose {
-}
-
-// Server -> client, terminates the call. Trailing metadata rides as __trailer__.
-// prefixed delivery headers, as in unary.
-message Trailers {
-  uint32 status = 1; // gRPC status code
-  string status_message = 2;
-}
-
-// Either direction, abnormal termination.
-message Cancel {
-  uint32 status = 1; // typically CANCELLED or DEADLINE_EXCEEDED
-  string reason = 2;
-}
