+++ vertx-grpc-eventbus/src/main/proto/io/vertx/grpc/eventbus/transport/v1alpha/eventbus_transport.proto
@@ -1,0 +1,68 @@
+syntax = "proto3";
+
+package io.vertx.grpc.eventbus.transport.v1alpha;
+
+option java_package = "io.vertx.grpc.eventbus.transport.v1alpha";
+option java_multiple_files = true;
+
+// A frame exchanged over the event bus during a streaming gRPC call. Unary calls
+// do not use these frames; they map to a plain request/reply. A streaming call
+// opens with a request whose reply is an Ack, after which each direction carries
+// the frames below over its own dedicated address. Frames are serialized as
+// protobuf binary. The handshake and flow control are described in the module
+// README.
+message TransportFrame {
+  uint64 seq = 1; // per-call, monotonic, advances on Message frames only
+
+  oneof frame {
+    Ack ack = 2; // server -> client, the open reply
+    Message message = 3; // a message payload, either direction
+    WindowUpdate window_update = 4; // flow-control credit, either direction
+    HalfClose half_close = 5; // client -> server, end of the request stream
+    Trailers trailers = 6; // server -> client, terminates the call
+    Cancel cancel = 7; // either direction, abnormal termination
+    Headers headers = 8; // server -> client, response metadata, before the first message
+  }
+}
+
+// Server -> client, the reply to a streaming call's opening request, carrying the
+// server's inbound address. Response initial metadata does not ride here; it
+// follows as a Headers frame.
+message Ack {
+  string server_address = 1; // address for client -> server frames
+  uint32 initial_window = 2; // messages the server grants the client to send
+}
+
+// Server -> client, response initial metadata, ordered ahead of the first response
+// message. The metadata itself rides as __header__. prefixed delivery headers.
+message Headers {
+}
+
+// A message payload, either direction. The serialized message in the call's wire
+// format, carried verbatim.
+message Message {
+  bytes payload = 1;
+}
+
+// Flow-control credit, either direction: grants the peer delta more messages to
+// send. Counted in messages, after HTTP/2's WINDOW_UPDATE.
+message WindowUpdate {
+  uint32 delta = 1;
+}
+
+// Client -> server, end of the request stream (half close).
+message HalfClose {
+}
+
+// Server -> client, terminates the call. Trailing metadata rides as __trailer__.
+// prefixed delivery headers.
+message Trailers {
+  uint32 status = 1; // gRPC status code
+  string status_message = 2;
+}
+
+// Either direction, abnormal termination.
+message Cancel {
+  uint32 status = 1; // typically CANCELLED or DEADLINE_EXCEEDED
+  string reason = 2;
+}
